Add fallback reference to base language
Running the app on OS X 10.8 had it showing the japanese text because it
didn't consider English and Base to be the same but it found japanese
text before running out of prefered languages. By adding an en.lproj
and its respective references the app would run the right localization
text for english and japanese.