Fix crash when X11WindowDriver failed to get window properties
X11WindowDriver uses `XGetWindowPropertyRef` to get a reference
to the properties of the active window. If this fails, however,
the outer function findFocusedWindow does not return "NO" to
indicate failure, because the result of the inner block is
ignored.

This change makes sure findFocusedWindow returns "NO" if the
inner block returns "NO".